cache write support (readx)
- fix RAW hazard bus ($1=$imm is not an hazard)
- tag bug fix
- raw hazard info
- debug conf in sim_env

testing:
- example imem0
- test_read
- test_write

+++ sim/cache.c
@@ -44,7 +44,7 @@ uint16_t cache_tag_get(struct cache *p_cache, uint8_t idx)
 {
 	struct tsram *p_tsram = &p_cache->tsram;
 
-	return (uint8_t)p_tsram->block_info[idx].tag;
+	return (uint16_t)p_tsram->block_info[idx].tag;
 }
 
 void cache_tag_set(struct cache *p_cache, uint8_t idx, uint16_t tag)
